Organized Angular files
